+++ src/cl.lightbox/CHANGELOG.rst
@@ -4,9 +4,27 @@ Cl.Lightbox
 
 1.1.0
 -----
-- Change enableKeys to keys and keys to keyCodes
-- Internal API refactor (_triggerAPI)
-- Changed image loader to be in normal tempalte
+- Change "enableKeys" to "keys" and "keys" to "keyCodes"
+- Internal API refactor (_triggerAPI to _triggerEvent and triggerCallback)
+- Changed image loader to be in default tpl
+- Content does not need to inject loader anymore
+- Move more variables to initialization (_setup)
+- Refactor control handling and added options to disable
+- Renamed build to _setup to be mor consistent
+- Fixed dimmer jumpint error
+- Reset cursor when using modalClickable and modalClosable
+- Inverse forceLoad option to disable it using false
+- Renamed transition to easing to be aligned with jquery animation
+- Added correct use of duration and speed
+- The responsive option was ignored
+- Remove autoscale option, this can be handeled using css only
+- Remove scrolling option, this can be handeled using css only
+- Removed link from options.lang.errorMessage
 
 1.0.1
 -----
+
+
+1.0.0
+-----
+- Initial release